Skeleton Project Files
Very general layout for building on.

Logic for all elements in the GUI need to be programmed.

I've included just a few sample lines that will populate values into the
QTableView widget.  When you run the program you'll see what i'm talking
about.

I'll update this as a I make more additions.  I just wanted to go ahead
and get this up so you guys can start working on it.