input pop_up implementation
Implements a pop_up window when going under "tools" and selecting "add
entry." From this point the window will ask for a given set of data. All
of this data is saved into a struct that will eventually be passed into
the function that displays this information in the correct location.