Added transaction cell combobox
When user double clicks Transaction cell, it will turn into a combo box
with option available from the "categories" file.
upon making a choice and hitting enter, the new value will be placed in
the cell, the combo box will be removed, and the value will be written
in place of the old one in the txt file.